+++ CHANGES.md
@@ -1,11 +1,19 @@
-1.1 (05-Jan-2015):
-* xen: pure OCaml implementation of `Clock.gmtime`. 
+### 1.2.0 (2016-12-21)
 
-1.0.0 (07-Dec-2013):
+* import `V1.MCLOCK` and `V1.PCLOCK` from `mirage-types` under `mirage-clock`
+  and `mirage-clock-lwt`
+
+### 1.1.0 (2015-01-05):
+
+* xen: pure OCaml implementation of `Clock.gmtime`
+
+### 1.0.0 (2013-12-07):
+
 * Remove unnecessary cstruct dependency.
 * Install ocamlfind packages as `mirage-clock-xen` and `mirage-clock-unix`.
 * Fix META file descriptions.
 * Add Travis tests.
 
-0.9.9 (05-Dec-2013):
+### 0.9.9 (2013-12-05):
+
 * Initial public release, based on mirage/mirage-platform#0.9.8